Move Sse41 GetColSums and GetAbsDxColSums to SimdSse41ColSums.cpp
Extract Sse41::GetColSums and Sse41::GetAbsDxColSums from
SimdSse41Statistic.cpp into a dedicated SimdSse41ColSums.cpp, and
register the new file in the VS2022 Sse41 project.
