+++ prj/vs2022/Sse41.vcxproj
@@ -33,6 +33,7 @@
     <ClCompile Include="..\..\src\Simd\SimdSse41BgrToRgb.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41BgrToYuvV2.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Binarization.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdSse41ColSums.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Conditional.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Cpu.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Crc32.cpp" />
